+++ src/entities/comment/ui/index.ts
@@ -1,0 +1,2 @@
+export * from "./CommentListSkeleton"
+export * from "./CommentList"
