+++ src/entities/user/model/hooks/queries/index.ts
@@ -1,1 +1,2 @@
 export * from "./useUsersQuery"
+export * from "./useUserQuery"
